ISSUE #4155 - move types into types folder
