+++ frontend/src/v5/store/tickets/groups/ticketGroups.types.ts
@@ -19,10 +19,31 @@ export interface ITicketGroupFromApi {
 	groupsByTicketId: Record<string, IGroupFromApi>,
 }
 
+export const OPERATIONS_TYPES = {
+	EXISTS: 'field',
+	NOT_EXISTS: 'field',
+	IS: 'text',
+	IS_NOT: 'text',
+	CONTAINS: 'text',
+	NOT_CONTAINS: 'text',
+	REGEX: 'regex',
+	EQUALS: 'number',
+	NOT_EQUALS: 'number',
+	GT: 'numberComparison',
+	GTE: 'numberComparison',
+	LT: 'numberComparison',
+	LTE: 'numberComparison',
+	IN_RANGE: 'numberRange',
+	NOT_IN_RANGE: 'numberRange',
+} as const;
+
+export type Operation = keyof typeof OPERATIONS_TYPES;
+export type OperationType = typeof OPERATIONS_TYPES[Operation];
+
 export interface IGroupRule {
 	field: string,
-	operator: string,
-	values: string[],
+	operation: Operation,
+	values?: (number | string)[],
 }
 
 export interface IGroupFromApi {
